Update app.py
get_all_rrsets was returning data when querying for a record that does not exist.  This would cause the DELETE to fail as it did not exist.  This simple change checks to make sure that the record name returned from get_all_rrsets is the same as the record that is being modified.